# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Public API Example Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "roger.medico@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/public-api/public/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/pettype": {
            "get": {
                "tags": [
                    "Pet Types"
                ],
                "summary": "Get list of pet types",
                "description": "Returns list of pet types",
                "operationId": "getPetTypeList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PetTypesResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Pet Types"
                ],
                "summary": "Store new pet type",
                "description": "Returns the pet type created data",
                "operationId": "storePetType",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StorePetTypeRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PetType"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "name": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "The name has already been taken."
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/pettype/{id}": {
            "get": {
                "tags": [
                    "Pet Types"
                ],
                "summary": "Get pet type information",
                "description": "Returns pet type data",
                "operationId": "getPetTypeById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "pet type id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PetTypeResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Object not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Pet Types"
                ],
                "summary": "Update existing pet type",
                "description": "Returns updated pet type data",
                "operationId": "updatePetType",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Pet type id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdatePetTypeRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PetType"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "name": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "The name must not be greater than 255 characters."
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Object not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Forbidden"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Pet Types"
                ],
                "summary": "Delete existing pet type",
                "description": "Deletes a pet type record and returns no content",
                "operationId": "deletePetType",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Pet type id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Resource Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Object not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Forbidden"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "PetType": {
                "title": "PetType",
                "description": "Pet type model",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "id",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "title": "name",
                        "description": "name",
                        "type": "string",
                        "format": "string",
                        "example": "cat"
                    },
                    "token_id": {
                        "title": "token_id",
                        "description": "token id identifying the author of the request",
                        "type": "integer",
                        "format": "int64",
                        "example": "10"
                    },
                    "created_at": {
                        "title": "created_at",
                        "description": "when record was created",
                        "type": "string",
                        "format": "datetime",
                        "example": "2021-12-10T10:31:23.000000Z"
                    },
                    "updated_at": {
                        "title": "updated_at",
                        "description": "when record was updated",
                        "type": "string",
                        "format": "datetime",
                        "example": "2021-12-10T10:31:23.000000Z"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "PetType"
                }
            },
...
            "StorePetTypeRequest": {
                "title": "StorePetTypeRequest",
                "description": "Store Pet Type request body data",
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "Name of the new pet type",
                        "type": "string",
                        "example": "cat"
                    }
                },
                "type": "object"
            },
...
            "UpdatePetTypeRequest": {
                "title": "UpdatePetTypeRequest",
                "description": "Update Pet Type request body data",
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "Updated ame for pet type",
                        "type": "string",
                        "example": "Lion"
                    }
                },
                "type": "object"
            },
...
            "PetTypeResource": {
                "title": "PetTypeResource",
                "description": "Pet type resource",
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/PetType"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "PetTypeResource"
                }
            },
            "PetTypesResource": {
                "title": "PetTypesResource",
                "description": "Pet types resource",
                "properties": {
                    "data": {
                        "title": "Data",
                        "description": "Data wrapper",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PetType"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "PetTypesResource"
                }
            },
...
        },
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        []
    ]
}